requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
llama-cpp-python>=0.2.90
huggingface_hub>=0.23.0
pydantic>=2.0.0
